Add Source commands to INBC (#440)
