+++ inbc-program/inbc/command/command_factory.py
@@ -10,9 +10,10 @@ from .command import Command, RestartCommand, QueryCommand
 from .ota_command import FotaCommand, SotaCommand, PotaCommand, AotaCommand
 from .config_command import GetConfigCommand, SetConfigCommand, LoadConfigCommand, AppendConfigCommand, \
     RemoveConfigCommand
+from .source_command import SourceCommand
 
 from inbm_common_lib.constants import CONFIG_LOAD, CONFIG_APPEND, CONFIG_REMOVE
-from inbm_lib.constants import AOTA, FOTA, SOTA, POTA, RESTART, QUERY
+from inbm_lib.constants import AOTA, FOTA, SOTA, POTA, RESTART, QUERY, SOURCE
 
 
 def create_command_factory(cmd: str, broker: IBroker) -> Command:
@@ -44,5 +45,7 @@ def create_command_factory(cmd: str, broker: IBroker) -> Command:
         return AppendConfigCommand(broker)
     if cmd == CONFIG_REMOVE:
         return RemoveConfigCommand(broker)
+    if cmd == SOURCE:
+        return SourceCommand(broker)
 
     raise InbcException(f"Unsupported command {cmd}")
