+++ inbc-program/inbc/constants.py
@@ -16,6 +16,7 @@ MQTT_HOST = 'localhost'
 
 # Timer
 MAX_TIME_LIMIT = 120
+SOURCE_TIME_LIMIT = 30
 FOTA_TIME_LIMIT = 260
 AOTA_TIME_LIMIT = 500
 SOTA_TIME_LIMIT = 900
