+++ inbc-program/inbc/parser/__init__.py
@@ -1,0 +1,4 @@
+"""
+    Copyright (C) 2017-2023 Intel Corporation
+    SPDX-License-Identifier: Apache-2.0
+"""
