+++ inbc-program/pyproject.toml
@@ -13,17 +13,17 @@ max_line_length = 99
 
 # complexity; see https://wemake-python-stylegui.de/en/latest/pages/usage/violations/complexity.html#complexity
 max_jones_score = 12
-max_cognitive_score = 208 # extremely high
-max_line_complexity = 21 # high
-max_imports = 49 # very high
+max_cognitive_score = 150 # extremely high
+max_line_complexity = 18 # high
+max_imports = 20 # very high
 max_module_members = 14 # very high
 max_imported_names = 62 # very high
-max_module_expressions = 47 # very high
+max_module_expressions = 30 # very high
 max_function_expressions = 8 # high
 max_local_variables = 16 # very high
 max_string_usages = 24 # extremely high
-max_arguments = 11 # very high
-max_returns = 11 # very high
+max_arguments = 7 # very high
+max_returns = 12 # very high
 max_expressions = 26 # very high
 max_methods = 30 # very high
 max_base_classes = 3
