+++ inbc-program/tests/unit/test_utility.py
@@ -7,9 +7,9 @@ class TestUtility(TestCase):
     def test_search_keyword_true(self) -> None:
         payload = 'Status message FAILED'
         output = search_keyword(payload, ["Configuration", "command", "FAILED"])
-        assert output == True
+        self.assertEqual(output, True)
 
     def test_search_keyword_false(self) -> None:
         payload = 'Status message SUCCESSFUL'
         output = search_keyword(payload, ["Commands"])
-        assert output == False
+        self.assertEqual(output, False)
