+++ inbm-lib/pyproject.toml
@@ -7,10 +7,8 @@ doctests = false
 enable_extensions = "G"
 
 exclude = ["tests",
-    "vision/tests",
     "inbm_lib/tests",
     "inbm_common_lib/tests",
-    "node/tests",
     "mqttclient/tests"]
 
 accept_encodings = "utf-8"
